Fixed bug where reset windows would always happen even if no was pressed.
git-svn-id: https://192.168.3.139:8443/svn/Medical/trunk@4049 956a8c8d-885e-cb46-a0b9-a2f355b4c364

+++ Standalone/GUI/Options/OptionsDialog.cs
@@ -265,16 +265,19 @@ namespace Medical.GUI
         {
             MessageBox.show("This will reset all window positions back to their defaults.\nIt will remove all icons from the Task Bar and you will have to put them back.\nAnomalous Medical will need to restart and you will lose any unsaved data.\nAre you sure you want to continue?", "Restart", MessageBoxStyle.IconInfo | MessageBoxStyle.Yes | MessageBoxStyle.No, delegate(MessageBoxStyle result)
             {
-                if (guiManager.deleteWindowsFile())
+                if (result == MessageBoxStyle.Yes)
                 {
-                    if (RequestRestart != null)
+                    if (guiManager.deleteWindowsFile())
                     {
-                        RequestRestart.Invoke(this, EventArgs.Empty);
+                        if (RequestRestart != null)
+                        {
+                            RequestRestart.Invoke(this, EventArgs.Empty);
+                        }
                     }
-                }
-                else
-                {
-                    MessageBox.show(String.Format("Could not delete windows.ini file located at '{0}'.\nPlease delete this file manually to reset your UI.", MedicalConfig.WindowsFile), "Error", MessageBoxStyle.Ok | MessageBoxStyle.IconError);
+                    else
+                    {
+                        MessageBox.show(String.Format("Could not delete windows.ini file located at '{0}'.\nPlease delete this file manually to reset your UI.", MedicalConfig.WindowsFile), "Error", MessageBoxStyle.Ok | MessageBoxStyle.IconError);
+                    }
                 }
             });
         }
